update 1# change max value Fibonacci return

+++ LemonWayTest.ServiceTests/FibonacciServiceTests.cs
@@ -1,11 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using LemonWayTest.Service;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Data;
 
 namespace LemonWayTest.Service.Tests
 {
@@ -21,11 +14,11 @@ namespace LemonWayTest.Service.Tests
         {
             // Arrange
             int n = (int)TestContext.DataRow[0];
-            int expected = (int)TestContext.DataRow[1];
+            decimal expected = (decimal)TestContext.DataRow[1];
             string message = TestContext.DataRow[2].ToString();
 
             // Act
-            int result = ServiceFactory.Current.FibonacciService.GetNthFibonacci(n);
+            decimal result = ServiceFactory.Current.FibonacciService.GetNthFibonacci(n);
 
             // Assert
             Assert.AreEqual(expected, result,message);
